+++ backend/.gitignore
@@ -1,1 +1,1 @@
-backend/node_modules
+node_modules
