+++ backend/docker-compose.yml
@@ -1,0 +1,11 @@
+version: "3.9"
+services:
+  postgres:
+    image: postgres:13
+    environment:
+      POSTGRES_USER: postgres
+      POSTGRES_PASSWORD: postgres
+      POSTGRES_DB: movies-db
+    command: -c fsync=off -c full_page_writes=off -c synchronous_commit=off -c max_connections=500
+    ports:
+      - 10002:5432
