--- backend/movies-db.ts
@@ -1,485 +1,0 @@
-[
-    {
-        "title": "The Shawshank Redemption",
-        "description": "Two imprisoned men bond over a number of years, finding solace and eventual redemption through acts of common decency.",
-        "year": 1994,
-        "duration": 142,
-        "picture": "",
-        "favorite": false,
-        "rating": 9.3,
-        "tags": []
-    },
-    {
-        "title": "The Godfather",
-        "description": "The aging patriarch of an organized crime dynasty transfers control of his clandestine empire to his reluctant son.",
-        "year": 1972,
-        "duration": 175,
-        "picture": "",
-        "favorite": false,
-        "rating": 9.2,
-        "tags": []
-    },
-    {
-        "title": "The Dark Knight",
-        "description": "When the menace known as the Joker emerges from his mysterious past, he wreaks havoc and chaos on the people of Gotham.",
-        "year": 2008,
-        "duration": 152,
-        "picture": "",
-        "favorite": false,
-        "rating": 9.0,
-        "tags": []
-    },
-    {
-        "title": "12 Angry Men",
-        "description": "A jury holdout attempts to prevent a miscarriage of justice by forcing his colleagues to reconsider the evidence.",
-        "year": 1957,
-        "duration": 96,
-        "picture": "",
-        "favorite": false,
-        "rating": 9.0,
-        "tags": []
-    },
-    {
-        "title": "Schindler's List",
-        "description": "In German-occupied Poland during World War II, industrialist Oskar Schindler gradually becomes concerned for his Jewish workforce after witnessing their persecution by the Nazis.",
-        "year": 1993,
-        "duration": 195,
-        "picture": "",
-        "favorite": false,
-        "rating": 9.0,
-        "tags": []
-    },
-    {
-        "title": "Pulp Fiction",
-        "description": "The lives of two mob hitmen, a boxer, a gangster and his wife, and a pair of diner bandits intertwine in four tales of violence and redemption.",
-        "year": 1994,
-        "duration": 154,
-        "picture": "",
-        "favorite": false,
-        "rating": 8.9,
-        "tags": []
-    },
-    {
-        "title": "The Lord of the Rings: The Return of the King",
-        "description": "Gandalf and Aragorn lead the World of Men against Sauron's army to draw his gaze from Frodo and Sam as they approach Mount Doom with the One Ring.",
-        "year": 2003,
-        "duration": 201,
-        "picture": "",
-        "favorite": false,
-        "rating": 8.9,
-        "tags": []
-    },
-    {
-        "title": "Fight Club",
-        "description": "An insomniac office worker and a devil-may-care soap maker form an underground fight club that evolves into much more.",
-        "year": 1999,
-        "duration": 139,
-        "picture": "",
-        "favorite": false,
-        "rating": 8.8,
-        "tags": []
-    },
-    {
-        "title": "Forrest Gump",
-        "description": "The presidencies of Kennedy and Johnson, the Vietnam War, the Watergate scandal and other historical events unfold from the perspective of an Alabama man with an IQ of 75.",
-        "year": 1994,
-        "duration": 142,
-        "picture": "",
-        "favorite": false,
-        "rating": 8.8,
-        "tags": []
-    },
-    {
-        "title": "Inception",
-        "description": "A thief who steals corporate secrets through the use of dream-sharing technology is given the inverse task of planting an idea into the mind of a C.E.O.",
-        "year": 2010,
-        "duration": 148,
-        "picture": "",
-        "favorite": false,
-        "rating": 8.8,
-        "tags": []
-    },
-    {
-        "title": "The Matrix",
-        "description": "A computer hacker learns from mysterious rebels about the true nature of his reality and his role in the war against its controllers.",
-        "year": 1999,
-        "duration": 136,
-        "picture": "",
-        "favorite": false,
-        "rating": 8.7,
-        "tags": []
-    },
-    {
-        "title": "Goodfellas",
-        "description": "The story of Henry Hill and his life in the mob, covering his relationship with his wife Karen Hill and his mob partners.",
-        "year": 1990,
-        "duration": 146,
-        "picture": "",
-        "favorite": false,
-        "rating": 8.7,
-        "tags": []
-    },
-    {
-        "title": "The Silence of the Lambs",
-        "description": "A young F.B.I. cadet must receive the help of an incarcerated and manipulative cannibal killer to catch another serial killer, a madman who skins his victims.",
-        "year": 1991,
-        "duration": 118,
-        "picture": "",
-        "favorite": false,
-        "rating": 8.6,
-        "tags": []
-    },
-    {
-        "title": "Interstellar",
-        "description": "A team of explorers travel through a wormhole in space in an attempt to ensure humanity's survival.",
-        "year": 2014,
-        "duration": 169,
-        "picture": "",
-        "favorite": false,
-        "rating": 8.6,
-        "tags": []
-    },
-    {
-        "title": "Saving Private Ryan",
-        "description": "Following the Normandy Landings, a group of U.S. soldiers go behind enemy lines to retrieve a paratrooper whose brothers have been killed in action.",
-        "year": 1998,
-        "duration": 169,
-        "picture": "",
-        "favorite": false,
-        "rating": 8.6,
-        "tags": []
-    },
-    {
-        "title": "The Green Mile",
-        "description": "The lives of guards on Death Row are affected by one of their charges: a black man accused of child murder and rape, yet who has a mysterious gift.",
-        "year": 1999,
-        "duration": 189,
-        "picture": "",
-        "favorite": false,
-        "rating": 8.6,
-        "tags": []
-    },
-    {
-        "title": "Se7en",
-        "description": "Two detectives, a rookie and a veteran, hunt a serial killer who uses the seven deadly sins as his motives.",
-        "year": 1995,
-        "duration": 127,
-        "picture": "",
-        "favorite": false,
-        "rating": 8.6,
-        "tags": []
-    },
-    {
-        "title": "The Usual Suspects",
-        "description": "A sole survivor tells of the twisty events leading up to a horrific gun battle on a boat, which began when five criminals met at a seemingly random police lineup.",
-        "year": 1995,
-        "duration": 106,
-        "picture": "",
-        "favorite": false,
-        "rating": 8.5,
-        "tags": []
-    },
-    {
-        "title": "The Lion King",
-        "description": "Lion prince Simba and his father are targeted by his bitter uncle, who wants to ascend the throne himself.",
-        "year": 1994,
-        "duration": 88,
-        "picture": "",
-        "favorite": false,
-        "rating": 8.5,
-        "tags": []
-    },
-    {
-        "title": "Star Wars: Episode V - The Empire Strikes Back",
-        "description": "After the Rebels are brutally overpowered by the Empire on the ice planet Hoth, Luke Skywalker begins Jedi training with Yoda, while his friends are pursued across the galaxy by Darth Vader.",
-        "year": 1980,
-        "duration": 124,
-        "picture": "",
-        "favorite": false,
-        "rating": 8.7,
-        "tags": []
-    },
-    [
-        {
-            "title": "The Pianist",
-            "description": "A Polish Jewish musician struggles to survive the destruction of the Warsaw ghetto of World War II.",
-            "year": 2002,
-            "duration": 150,
-            "picture": "",
-            "favorite": false,
-            "rating": 8.5,
-            "tags": []
-        },
-        {
-            "title": "Gladiator",
-            "description": "A former Roman General sets out to exact vengeance against the corrupt emperor who murdered his family and sent him into slavery.",
-            "year": 2000,
-            "duration": 155,
-            "picture": "",
-            "favorite": false,
-            "rating": 8.5,
-            "tags": []
-        },
-        {
-            "title": "The Departed",
-            "description": "An undercover cop and a mole in the police attempt to identify each other while infiltrating an Irish gang in South Boston.",
-            "year": 2006,
-            "duration": 151,
-            "picture": "",
-            "favorite": false,
-            "rating": 8.5,
-            "tags": []
-        },
-        {
-            "title": "The Prestige",
-            "description": "After a tragic accident, two stage magicians engage in a battle to create the ultimate illusion while sacrificing everything they have to outwit each other.",
-            "year": 2006,
-            "duration": 130,
-            "picture": "",
-            "favorite": false,
-            "rating": 8.5,
-            "tags": []
-        },
-        {
-            "title": "Memento",
-            "description": "A man with short-term memory loss attempts to track down his wife's murderer.",
-            "year": 2000,
-            "duration": 113,
-            "picture": "",
-            "favorite": false,
-            "rating": 8.4,
-            "tags": []
-        },
-        {
-            "title": "Whiplash",
-            "description": "A promising young drummer enrolls at a cutthroat music conservatory where his dreams of greatness are mentored by an instructor who will stop at nothing to realize a student's potential.",
-            "year": 2014,
-            "duration": 106,
-            "picture": "",
-            "favorite": false,
-            "rating": 8.5,
-            "tags": []
-        },
-        {
-            "title": "The Intouchables",
-            "description": "After he becomes a quadriplegic from a paragliding accident, an aristocrat hires a young man from the projects to be his caregiver.",
-            "year": 2011,
-            "duration": 112,
-            "picture": "",
-            "favorite": false,
-            "rating": 8.5,
-            "tags": []
-        },
-        {
-            "title": "Parasite",
-            "description": "Greed and class discrimination threaten the newly formed symbiotic relationship between the wealthy Park family and the destitute Kim clan.",
-            "year": 2019,
-            "duration": 132,
-            "picture": "",
-            "favorite": false,
-            "rating": 8.6,
-            "tags": []
-        },
-        {
-            "title": "The Shining",
-            "description": "A family heads to an isolated hotel for the winter where a sinister presence influences the father into violence, while his psychic son sees horrific forebodings from both past and future.",
-            "year": 1980,
-            "duration": 146,
-            "picture": "",
-            "favorite": false,
-            "rating": 8.4,
-            "tags": []
-        },
-        {
-            "title": "American Beauty",
-            "description": "A sexually frustrated suburban father has a mid-life crisis after becoming infatuated with his daughter's best friend.",
-            "year": 1999,
-            "duration": 122,
-            "picture": "",
-            "favorite": false,
-            "rating": 8.3,
-            "tags": []
-        },
-        {
-            "title": "A Beautiful Mind",
-            "description": "After John Nash, a brilliant but asocial mathematician, accepts secret work in cryptography, his life takes a turn for the nightmarish.",
-            "year": 2001,
-            "duration": 135,
-            "picture": "",
-            "favorite": false,
-            "rating": 8.2,
-            "tags": []
-        },
-        {
-            "title": "The Wolf of Wall Street",
-            "description": "Based on the true story of Jordan Belfort, from his rise to a wealthy stock-broker living the high life to his fall involving crime, corruption and the federal government.",
-            "year": 2013,
-            "duration": 180,
-            "picture": "",
-            "favorite": false,
-            "rating": 8.2,
-            "tags": []
-        },
-        {
-            "title": "Mad Max: Fury Road",
-            "description": "In a post-apocalyptic wasteland, a woman rebels against a tyrannical ruler in search for her homeland with the aid of a group of female prisoners, a psychotic worshiper, and a drifter named Max.",
-            "year": 2015,
-            "duration": 120,
-            "picture": "",
-            "favorite": false,
-            "rating": 8.1,
-            "tags": []
-        },
-        {
-            "title": "The Social Network",
-            "description": "Harvard student Mark Zuckerberg creates the social networking site that would become known as Facebook, but is later sued by two brothers who claimed he stole their idea, and the co-founder who was later squeezed out of the business.",
-            "year": 2010,
-            "duration": 120,
-            "picture": "",
-            "favorite": false,
-            "rating": 7.7,
-            "tags": []
-        },
-        {
-            "title": "Her",
-            "description": "In a near future, a lonely writer develops an unlikely relationship with an operating system designed to meet his every need.",
-            "year": 2013,
-            "duration": 126,
-            "picture": "",
-            "favorite": false,
-            "rating": 8.0,
-            "tags": []
-        },
-        {
-            "title": "Joker",
-            "description": "In Gotham City's fractured society, a mentally troubled comedian embarks on a downward spiral of revolution and bloody crime.",
-            "year": 2019,
-            "duration": 122,
-            "picture": "",
-            "favorite": false,
-            "rating": 8.4,
-            "tags": []
-        },
-        {
-            "title": "1917",
-            "description": "Two soldiers are tasked with crossing enemy lines to deliver a message that will stop a deadly attack during World War I.",
-            "year": 2019,
-            "duration": 119,
-            "picture": "",
-            "favorite": false,
-            "rating": 8.3,
-            "tags": []
-        },
-        {
-            "title": "The Grand Budapest Hotel",
-            "description": "A writer encounters the owner of an aging high-class hotel, who tells him of his early years serving as a lobby boy in the hotel's glorious years under an exceptional concierge.",
-            "year": 2014,
-            "duration": 99,
-            "picture": "",
-            "favorite": false,
-            "rating": 8.1,
-            "tags": []
-        },
-        {
-            "title": "Django Unchained",
-            "description": "With the help of a German bounty-hunter, a freed slave sets out to rescue his wife from a brutal Mississippi plantation owner.",
-            "year": 2012,
-            "duration": 165,
-            "picture": "",
-            "favorite": false,
-            "rating": 8.4,
-            "tags": []
-        },
-        {
-            "title": "Logan",
-            "description": "In the near future, a weary Logan cares for an ailing Professor X somewhere on the Mexican border. However, Logan's attempts to hide from the world, and his legacy, are upended when a young mutant arrives, pursued by dark forces.",
-            "year": 2017,
-            "duration": 137,
-            "picture": "",
-            "favorite": false,
-            "rating": 8.1,
-            "tags": []
-        },
-        {
-            "title": "The Truman Show",
-            "description": "An insurance salesman discovers his whole life is actually a reality TV show.",
-            "year": 1998,
-            "duration": 103,
-            "picture": "",
-            "favorite": false,
-            "rating": 8.2,
-            "tags": []
-        },
-        {
-            "title": "Shutter Island",
-            "description": "In 1954, a U.S. Marshal investigates the disappearance of a murderer who escaped from a hospital for the criminally insane.",
-            "year": 2010,
-            "duration": 138,
-            "picture": "",
-            "favorite": false,
-            "rating": 8.2,
-            "tags": []
-        },
-        {
-            "title": "A Star Is Born",
-            "description": "A musician helps a young singer find fame as age and alcoholism send his own career into a downward spiral.",
-            "year": 2018,
-            "duration": 136,
-            "picture": "",
-            "favorite": false,
-            "rating": 7.6,
-            "tags": []
-        },
-        {
-            "title": "Blade Runner 2049",
-            "description": "Young Blade Runner K's discovery of a long-buried secret leads him to track down former Blade Runner Rick Deckard, who's been missing for thirty years.",
-            "year": 2017,
-            "duration": 164,
-            "picture": "",
-            "favorite": false,
-            "rating": 8.0,
-            "tags": []
-        },
-        {
-            "title": "La La Land",
-            "description": "While navigating their careers in Los Angeles, a pianist and an actress fall in love while attempting to reconcile their aspirations for the future.",
-            "year": 2016,
-            "duration": 128,
-            "picture": "",
-            "favorite": false,
-            "rating": 8.0,
-            "tags": []
-        },
-        {
-            "title": "Once Upon a Time in Hollywood",
-            "description": "A faded television actor and his stunt double strive to achieve fame and success in the final years of Hollywood's Golden Age in 1969 Los Angeles.",
-            "year": 2019,
-            "duration": 161,
-            "picture": "",
-            "favorite": false,
-            "rating": 7.6,
-            "tags": []
-        },
-        {
-            "title": "Jojo Rabbit",
-            "description": "A young boy in Hitler's army finds out his mother is hiding a Jewish girl in their home.",
-            "year": 2019,
-            "duration": 108,
-            "picture": "",
-            "favorite": false,
-            "rating": 7.9,
-            "tags": []
-        },
-        {
-            "title": "Tenet",
-            "description": "Armed with only one word, Tenet, and fighting for the survival of the world, a protagonist journeys through a twilight world of international espionage on a mission that will unfold in something beyond real time.",
-            "year": 2020,
-            "duration": 150,
-            "picture": "",
-            "favorite": false,
-            "rating": 7.4,
-            "tags": []
-        }
-    ]
-
-]
