+++ backend/package.json
@@ -9,10 +9,12 @@
   "license": "ISC",
   "description": "",
   "dependencies": {
+    "@prisma/client": "^5.20.0",
     "@types/express": "^5.0.0",
     "@types/node": "^22.7.4",
     "express": "^4.21.0",
     "node": "^20.18.0",
+    "prisma": "^5.20.0",
     "typescript": "^5.6.2"
   },
   "devDependencies": {
