+++ backend/src/movies-db.ts
@@ -1,0 +1,482 @@
+export const movies = [
+    {
+        "img": "",
+        "title": "The Shawshank Redemption",
+        "favorite": false,
+        "year": 1994,
+        "duration": 142,
+        "rating": 9.3,
+        "description": "Two imprisoned men bond over a number of years, finding solace and eventual redemption through acts of common decency.",
+        "tags": []
+    },
+    {
+        "img": "",
+        "title": "The Godfather",
+        "favorite": false,
+        "year": 1972,
+        "duration": 175,
+        "rating": 9.2,
+        "description": "The aging patriarch of an organized crime dynasty transfers control of his clandestine empire to his reluctant son.",
+        "tags": []
+    },
+    {
+        "img": "",
+        "title": "The Dark Knight",
+        "favorite": false,
+        "year": 2008,
+        "duration": 152,
+        "rating": 9.0,
+        "description": "When the menace known as the Joker emerges from his mysterious past, he wreaks havoc and chaos on the people of Gotham.",
+        "tags": []
+    },
+    {
+        "img": "",
+        "title": "12 Angry Men",
+        "favorite": false,
+        "year": 1957,
+        "duration": 96,
+        "rating": 9.0,
+        "description": "A jury holdout attempts to prevent a miscarriage of justice by forcing his colleagues to reconsider the evidence.",
+        "tags": []
+    },
+    {
+        "img": "",
+        "title": "Schindler's List",
+        "favorite": false,
+        "year": 1993,
+        "duration": 195,
+        "rating": 9.0,
+        "description": "In German-occupied Poland during World War II, industrialist Oskar Schindler gradually becomes concerned for his Jewish workforce after witnessing their persecution by the Nazis.",
+        "tags": []
+    },
+    {
+        "img": "",
+        "title": "Pulp Fiction",
+        "favorite": false,
+        "year": 1994,
+        "duration": 154,
+        "rating": 8.9,
+        "description": "The lives of two mob hitmen, a boxer, a gangster and his wife, and a pair of diner bandits intertwine in four tales of violence and redemption.",
+        "tags": []
+    },
+    {
+        "img": "",
+        "title": "The Lord of the Rings: The Return of the King",
+        "favorite": false,
+        "year": 2003,
+        "duration": 201,
+        "rating": 8.9,
+        "description": "Gandalf and Aragorn lead the World of Men against Sauron's army to draw his gaze from Frodo and Sam as they approach Mount Doom with the One Ring.",
+        "tags": []
+    },
+    {
+        "img": "",
+        "title": "Fight Club",
+        "favorite": false,
+        "year": 1999,
+        "duration": 139,
+        "rating": 8.8,
+        "description": "An insomniac office worker and a devil-may-care soap maker form an underground fight club that evolves into much more.",
+        "tags": []
+    },
+    {
+        "img": "",
+        "title": "Forrest Gump",
+        "favorite": false,
+        "year": 1994,
+        "duration": 142,
+        "rating": 8.8,
+        "description": "The presidencies of Kennedy and Johnson, the Vietnam War, the Watergate scandal and other historical events unfold from the perspective of an Alabama man with an IQ of 75.",
+        "tags": []
+    },
+    {
+        "img": "",
+        "title": "Inception",
+        "favorite": false,
+        "year": 2010,
+        "duration": 148,
+        "rating": 8.8,
+        "description": "A thief who steals corporate secrets through the use of dream-sharing technology is given the inverse task of planting an idea into the mind of a C.E.O.",
+        "tags": []
+    },
+    {
+        "img": "",
+        "title": "The Matrix",
+        "favorite": false,
+        "year": 1999,
+        "duration": 136,
+        "rating": 8.7,
+        "description": "A computer hacker learns from mysterious rebels about the true nature of his reality and his role in the war against its controllers.",
+        "tags": []
+    },
+    {
+        "img": "",
+        "title": "Goodfellas",
+        "favorite": false,
+        "year": 1990,
+        "duration": 146,
+        "rating": 8.7,
+        "description": "The story of Henry Hill and his life in the mob, covering his relationship with his wife Karen Hill and his mob partners.",
+        "tags": []
+    },
+    {
+        "img": "",
+        "title": "The Silence of the Lambs",
+        "favorite": false,
+        "year": 1991,
+        "duration": 118,
+        "rating": 8.6,
+        "description": "A young F.B.I. cadet must receive the help of an incarcerated and manipulative cannibal killer to catch another serial killer, a madman who skins his victims.",
+        "tags": []
+    },
+    {
+        "img": "",
+        "title": "Interstellar",
+        "favorite": false,
+        "year": 2014,
+        "duration": 169,
+        "rating": 8.6,
+        "description": "A team of explorers travel through a wormhole in space in an attempt to ensure humanity's survival.",
+        "tags": []
+    },
+    {
+        "img": "",
+        "title": "Saving Private Ryan",
+        "favorite": false,
+        "year": 1998,
+        "duration": 169,
+        "rating": 8.6,
+        "description": "Following the Normandy Landings, a group of U.S. soldiers go behind enemy lines to retrieve a paratrooper whose brothers have been killed in action.",
+        "tags": []
+    },
+    {
+        "img": "",
+        "title": "The Green Mile",
+        "favorite": false,
+        "year": 1999,
+        "duration": 189,
+        "rating": 8.6,
+        "description": "The lives of guards on Death Row are affected by one of their charges: a black man accused of child murder and rape, yet who has a mysterious gift.",
+        "tags": []
+    },
+    {
+        "img": "",
+        "title": "Se7en",
+        "favorite": false,
+        "year": 1995,
+        "duration": 127,
+        "rating": 8.6,
+        "description": "Two detectives, a rookie and a veteran, hunt a serial killer who uses the seven deadly sins as his motives.",
+        "tags": []
+    },
+    {
+        "img": "",
+        "title": "The Usual Suspects",
+        "favorite": false,
+        "year": 1995,
+        "duration": 106,
+        "rating": 8.5,
+        "description": "A sole survivor tells of the twisty events leading up to a horrific gun battle on a boat, which began when five criminals met at a seemingly random police lineup.",
+        "tags": []
+    },
+    {
+        "img": "",
+        "title": "The Lion King",
+        "favorite": false,
+        "year": 1994,
+        "duration": 88,
+        "rating": 8.5,
+        "description": "Lion prince Simba and his father are targeted by his bitter uncle, who wants to ascend the throne himself.",
+        "tags": []
+    },
+    {
+        "img": "",
+        "title": "Star Wars: Episode V - The Empire Strikes Back",
+        "favorite": false,
+        "year": 1980,
+        "duration": 124,
+        "rating": 8.7,
+        "description": "After the Rebels are brutally overpowered by the Empire on the ice planet Hoth, Luke Skywalker begins Jedi training with Yoda, while his friends are pursued across the galaxy by Darth Vader.",
+        "tags": []
+    },
+    {
+        "img": "",
+        "title": "The Pianist",
+        "favorite": false,
+        "year": 2002,
+        "duration": 150,
+        "rating": 8.5,
+        "description": "A Polish Jewish musician struggles to survive the destruction of the Warsaw ghetto of World War II.",
+        "tags": []
+    },
+    {
+        "img": "",
+        "title": "Gladiator",
+        "favorite": false,
+        "year": 2000,
+        "duration": 155,
+        "rating": 8.5,
+        "description": "A former Roman General sets out to exact vengeance against the corrupt emperor who murdered his family and sent him into slavery.",
+        "tags": []
+    },
+    {
+        "img": "",
+        "title": "The Departed",
+        "favorite": false,
+        "year": 2006,
+        "duration": 151,
+        "rating": 8.5,
+        "description": "An undercover cop and a mole in the police attempt to identify each other while infiltrating an Irish gang in South Boston.",
+        "tags": []
+    },
+    {
+        "img": "",
+        "title": "The Prestige",
+        "favorite": false,
+        "year": 2006,
+        "duration": 130,
+        "rating": 8.5,
+        "description": "After a tragic accident, two stage magicians engage in a battle to create the ultimate illusion while sacrificing everything they have to outwit each other.",
+        "tags": []
+    },
+    {
+        "img": "",
+        "title": "Memento",
+        "favorite": false,
+        "year": 2000,
+        "duration": 113,
+        "rating": 8.4,
+        "description": "A man with short-term memory loss attempts to track down his wife's murderer.",
+        "tags": []
+    },
+    {
+        "img": "",
+        "title": "Whiplash",
+        "favorite": false,
+        "year": 2014,
+        "duration": 106,
+        "rating": 8.5,
+        "description": "A promising young drummer enrolls at a cutthroat music conservatory where his dreams of greatness are mentored by an instructor who will stop at nothing to realize a student's potential.",
+        "tags": []
+    },
+    {
+        "img": "",
+        "title": "The Intouchables",
+        "favorite": false,
+        "year": 2011,
+        "duration": 112,
+        "rating": 8.5,
+        "description": "After he becomes a quadriplegic from a paragliding accident, an aristocrat hires a young man from the projects to be his caregiver.",
+        "tags": []
+    },
+    {
+        "img": "",
+        "title": "Parasite",
+        "favorite": false,
+        "year": 2019,
+        "duration": 132,
+        "rating": 8.6,
+        "description": "Greed and class discrimination threaten the newly formed symbiotic relationship between the wealthy Park family and the destitute Kim clan.",
+        "tags": []
+    },
+    {
+        "img": "",
+        "title": "The Shining",
+        "favorite": false,
+        "year": 1980,
+        "duration": 146,
+        "rating": 8.4,
+        "description": "A family heads to an isolated hotel for the winter where a sinister presence influences the father into violence, while his psychic son sees horrific forebodings from both past and future.",
+        "tags": []
+    },
+    {
+        "img": "",
+        "title": "American Beauty",
+        "favorite": false,
+        "year": 1999,
+        "duration": 122,
+        "rating": 8.3,
+        "description": "A sexually frustrated suburban father has a mid-life crisis after becoming infatuated with his daughter's best friend.",
+        "tags": []
+    },
+    {
+        "img": "",
+        "title": "A Beautiful Mind",
+        "favorite": false,
+        "year": 2001,
+        "duration": 135,
+        "rating": 8.2,
+        "description": "After John Nash, a brilliant but asocial mathematician, accepts secret work in cryptography, his life takes a turn for the nightmarish.",
+        "tags": []
+    },
+    {
+        "img": "",
+        "title": "The Wolf of Wall Street",
+        "favorite": false,
+        "year": 2013,
+        "duration": 180,
+        "rating": 8.2,
+        "description": "Based on the true story of Jordan Belfort, from his rise to a wealthy stock-broker living the high life to his fall involving crime, corruption and the federal government.",
+        "tags": []
+    },
+    {
+        "img": "",
+        "title": "Mad Max: Fury Road",
+        "favorite": false,
+        "year": 2015,
+        "duration": 120,
+        "rating": 8.1,
+        "description": "In a post-apocalyptic wasteland, a woman rebels against a tyrannical ruler in search for her homeland with the aid of a group of female prisoners, a psychotic worshiper, and a drifter named Max.",
+        "tags": []
+    },
+    {
+        "img": "",
+        "title": "The Social Network",
+        "favorite": false,
+        "year": 2010,
+        "duration": 120,
+        "rating": 7.7,
+        "description": "Harvard student Mark Zuckerberg creates the social networking site that would become known as Facebook, but is later sued by two brothers who claimed he stole their idea, and the co-founder who was later squeezed out of the business.",
+        "tags": []
+    },
+    {
+        "img": "",
+        "title": "Her",
+        "favorite": false,
+        "year": 2013,
+        "duration": 126,
+        "rating": 8.0,
+        "description": "In a near future, a lonely writer develops an unlikely relationship with an operating system designed to meet his every need.",
+        "tags": []
+    },
+    {
+        "img": "",
+        "title": "Joker",
+        "favorite": false,
+        "year": 2019,
+        "duration": 122,
+        "rating": 8.4,
+        "description": "In Gotham City's fractured society, a mentally troubled comedian embarks on a downward spiral of revolution and bloody crime.",
+        "tags": []
+    },
+    {
+        "img": "",
+        "title": "1917",
+        "favorite": false,
+        "year": 2019,
+        "duration": 119,
+        "rating": 8.3,
+        "description": "Two soldiers are tasked with crossing enemy lines to deliver a message that will stop a deadly attack during World War I.",
+        "tags": []
+    },
+    {
+        "img": "",
+        "title": "The Grand Budapest Hotel",
+        "favorite": false,
+        "year": 2014,
+        "duration": 99,
+        "rating": 8.1,
+        "description": "A writer encounters the owner of an aging high-class hotel, who tells him of his early years serving as a lobby boy in the hotel's glorious years under an exceptional concierge.",
+        "tags": []
+    },
+    {
+        "img": "",
+        "title": "Django Unchained",
+        "favorite": false,
+        "year": 2012,
+        "duration": 165,
+        "rating": 8.4,
+        "description": "With the help of a German bounty-hunter, a freed slave sets out to rescue his wife from a brutal Mississippi plantation owner.",
+        "tags": []
+    },
+    {
+        "img": "",
+        "title": "Logan",
+        "favorite": false,
+        "year": 2017,
+        "duration": 137,
+        "rating": 8.1,
+        "description": "In the near future, a weary Logan cares for an ailing Professor X somewhere on the Mexican border. However, Logan's attempts to hide from the world, and his legacy, are upended when a young mutant arrives, pursued by dark forces.",
+        "tags": []
+    },
+    {
+        "img": "",
+        "title": "The Truman Show",
+        "favorite": false,
+        "year": 1998,
+        "duration": 103,
+        "rating": 8.2,
+        "description": "An insurance salesman discovers his whole life is actually a reality TV show.",
+        "tags": []
+    },
+    {
+        "img": "",
+        "title": "Shutter Island",
+        "favorite": false,
+        "year": 2010,
+        "duration": 138,
+        "rating": 8.2,
+        "description": "In 1954, a U.S. Marshal investigates the disappearance of a murderer who escaped from a hospital for the criminally insane.",
+        "tags": []
+    },
+    {
+        "img": "",
+        "title": "A Star Is Born",
+        "favorite": false,
+        "year": 2018,
+        "duration": 136,
+        "rating": 7.6,
+        "description": "A musician helps a young singer find fame as age and alcoholism send his own career into a downward spiral.",
+        "tags": []
+    },
+    {
+        "img": "",
+        "title": "Blade Runner 2049",
+        "favorite": false,
+        "year": 2017,
+        "duration": 164,
+        "rating": 8.0,
+        "description": "Young Blade Runner K's discovery of a long-buried secret leads him to track down former Blade Runner Rick Deckard, who's been missing for thirty years.",
+        "tags": []
+    },
+    {
+        "img": "",
+        "title": "La La Land",
+        "favorite": false,
+        "year": 2016,
+        "duration": 128,
+        "rating": 8.0,
+        "description": "While navigating their careers in Los Angeles, a pianist and an actress fall in love while attempting to reconcile their aspirations for the future.",
+        "tags": []
+    },
+    {
+        "img": "",
+        "title": "Once Upon a Time in Hollywood",
+        "favorite": false,
+        "year": 2019,
+        "duration": 161,
+        "rating": 7.6,
+        "description": "A faded television actor and his stunt double strive to achieve fame and success in the final years of Hollywood's Golden Age in 1969 Los Angeles.",
+        "tags": []
+    },
+    {
+        "img": "",
+        "title": "Jojo Rabbit",
+        "favorite": false,
+        "year": 2019,
+        "duration": 108,
+        "rating": 7.9,
+        "description": "A young boy in Hitler's army finds out his mother is hiding a Jewish girl in their home.",
+        "tags": []
+    },
+    {
+        "img": "",
+        "title": "Tenet",
+        "favorite": false,
+        "year": 2020,
+        "duration": 150,
+        "rating": 7.4,
+        "description": "Armed with only one word, Tenet, and fighting for the survival of the world, a protagonist journeys through a twilight world of international espionage on a mission that will unfold in something beyond real time.",
+        "tags": []
+    }
+]
